Fixed file menu items not working when activated with the keyboard
Note: this pops up a warning banner instead of the file dialog in Firefox, requiring the user to allow pop-ups for the website. If this becomes an issue, the menu could instead focus the file input when the item is selected so that it responds to <enter> and <space> and opens the dialog.